Move ListTab to widgets
