+++ aptlygui/views/snapshot_tab.py
@@ -4,7 +4,7 @@ from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QStandardItem, QStandardItemModel
 from PyQt5.QtWidgets import (QComboBox, QGridLayout, QFormLayout, QPushButton, QWidget, QListView, QAbstractItemView, QVBoxLayout, QCheckBox, QLabel)
 
-from aptlygui.views.list_tab import ListTab
+from aptlygui.widgets.list_tab import ListTab
 from aptlygui.views.wait_dialog import WaitDialog
